put into LLM.py
